requirements.txt
openai-whisper
streamlit==1.24.0
streamlit-audiorecorder